ENG-578: Split out node, relayer and network
